requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
autopep8==1.5.4
backcall==0.2.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.2
filelock==3.0.12
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
virtualenv==20.4.2
wcwidth==0.2.5
